frontend/requirements.txt
gradio>=4.0.0
python-dotenv>=1.0.0
google-adk>=0.3.0
google-genai>=0.3.0
protobuf>=4.0.0
vertexai>=0.1.0
google-cloud-aiplatform>=1.38.0
